Split POD checks out of podUtils
