+++ packages/lib/pod/package.json
@@ -10,6 +10,11 @@
       "types": "./dist/types/src/index.d.ts",
       "import": "./dist/esm/src/index.js",
       "require": "./dist/cjs/src/index.js"
+    },
+    "./podChecks": {
+      "types": "./dist/types/src/podChecks.d.ts",
+      "import": "./dist/esm/src/podChecks.js",
+      "require": "./dist/cjs/src/podChecks.js"
     }
   },
   "files": [
@@ -45,6 +50,7 @@
     "@semaphore-protocol/identity": "^3.15.2",
     "@types/circomlibjs": "^0.1.6",
     "@types/chai": "^4.3.5",
+    "@types/json-bigint": "^1.0.1",
     "@types/mocha": "^10.0.1",
     "circomlibjs": "^0.1.7",
     "eslint": "^8.57.0",
