+++ packages/lib/pod/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./pod";
+export * from "./podChecks";
 export * from "./podContent";
 export * from "./podCrypto";
 export * from "./podTypes";
