+++ packages/lib/pod/src/podContent.ts
@@ -1,9 +1,8 @@
 import { LeanIMT, LeanIMTMerkleProof } from "@zk-kit/lean-imt";
+import { checkPODName, checkPODValue } from "./podChecks";
 import { podMerkleTreeHash, podNameHash, podValueHash } from "./podCrypto";
 import { PODEntries, PODName, PODValue } from "./podTypes";
 import {
-  checkPODName,
-  checkPODValue,
   cloneOptionalPODValue,
   clonePODValue,
   deserializePODEntries,
